requirements.txt
fastapi
uvicorn
openai-whisper
torch
transformers
pydantic
python-multipart 